bug fix and blog
