+++ app/global.css
@@ -2,6 +2,11 @@
 @tailwind components;
 @tailwind utilities;
 
+::-webkit-scrollbar {
+  width: 0px;
+  background: transparent; /* make scrollbar transparent */
+}
+
 ::selection {
   background-color: #47a3f3;
   color: #fefefe;
